+++ lib/src/tools/connection/connection_builder.dart
@@ -21,7 +21,15 @@ class ConnectionBuilder extends BlocBuilder<ConnectionBloc, Connection> {
     @Deprecated('This builder will be removed. Use "offline" instead.')
     WidgetBuilder? onOffline,
     WidgetBuilder? offline,
-  }) : super(
+  })  : assert(
+          !(onOnline != null && online != null),
+          'The onOnline and online builders should NOT be used together. The onOnline builder is deprecated and can be safely removed.',
+        ),
+        assert(
+          !(onOffline != null && offline != null),
+          'The onOffline and offline builders should NOT be used together. The onOffline builder is deprecated and can be safely removed.',
+        ),
+        super(
           builder: (BuildContext context, Connection state) {
             const none = SizedBox.shrink();
             switch (state) {